+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file. This projec
 to [Semantic Versioning] (http://semver.org/). For change log format,
 use [Keep a Changelog] (http://keepachangelog.com/).
 
+## [2.1.0] - In progress
+
+### Added
+
+- New method `Environment::getArgumentMultiple()` to get values for an argument multiple
+
 ## [2.0.0] - 2021-09-08
 
 No changes were introduced since the previous beta 2 release.